+++ .gitignore
@@ -19,5 +19,4 @@ cache-ovm/
 .openzeppelin
 .openzeppelin/
 randomCodeSnippets.js
-deploy/*
-deploy/
+
